fix(protocol): re-pin the 2026-07-28 wire to the frozen schema
The spec froze on 2026-07-28; our types were generated from the RC's
schema/draft/, which is 195 insertions / 91 deletions behind it. So
4.0.0-alpha.1 advertises the wire string "2026-07-28" while implementing
something that is not it. Codegen now reads schema/2026-07-28/.

Four things the freeze reversed, each of which we had implemented the RC
way:

- serverInfo left DiscoverResult and went back into the result's _meta
  under io.modelcontextprotocol/serverInfo. The stateless model has no
  initialize result, so there is nowhere else for identity to ride. The
  client reads both places — the legacy wire never sets _meta, so
  preferring it costs nothing.
- ttlMs and cacheScope became *required* on DiscoverResult, which is now
  a CacheableResult. CachePolicies grows a `discover` surface rather than
  hardcoding a constant; the default stays NO_CACHE, so behaviour is
  unchanged until someone configures it.
- SubscriptionsListenResult came back, and with it a defined ending for a
  listen stream. close_all now answers each live listen request with the
  envelope before clearing, so graceful teardown is distinguishable from
  a dropped connection. That makes it async, as does close_subscriptions.
- elicitationId and notifications/elicitation/complete were deleted from
  the draft entirely. Both are 2025-11-25-only now, so elicit_url mints an
  id only on the legacy path and notify_elicitation_complete is a no-op on
  a draft handle — sending either would be inventing protocol.

The spec-allocated error codes also moved: -32003/-32004 became
-32021/-32022, and header mismatch got -32020 (the RC had no such type;
-32001 was ours). They were literals in five places, so this adds
turbomcp_core::codes as the one owner — the next reallocation is a
one-line change instead of a grep. The client already tolerated both
allocations.

Upstream also fixed NumberSchema on both *stable* wires:
minimum/maximum/default were integer, now number. Option<i64> would have
rejected `minimum: 0.5` on 2025-11-25 and 2025-06-18 — a live bug, not a
draft-only one.

667 tests pass; clippy -D warnings clean.

+++ crates/turbomcp-protocol/src/v2025_06_18/types.rs
@@ -4026,10 +4026,10 @@ pub struct NotificationParams {
 ///      "type": "string"
 ///    },
 ///    "maximum": {
-///      "type": "integer"
+///      "type": "number"
 ///    },
 ///    "minimum": {
-///      "type": "integer"
+///      "type": "number"
 ///    },
 ///    "title": {
 ///      "type": "string"
@@ -4050,9 +4050,9 @@ pub struct NumberSchema {
     #[serde(default, skip_serializing_if = "::core::option::Option::is_none")]
     pub description: ::core::option::Option<::alloc::string::String>,
     #[serde(default, skip_serializing_if = "::core::option::Option::is_none")]
-    pub maximum: ::core::option::Option<i64>,
+    pub maximum: ::core::option::Option<f64>,
     #[serde(default, skip_serializing_if = "::core::option::Option::is_none")]
-    pub minimum: ::core::option::Option<i64>,
+    pub minimum: ::core::option::Option<f64>,
     #[serde(default, skip_serializing_if = "::core::option::Option::is_none")]
     pub title: ::core::option::Option<::alloc::string::String>,
     #[serde(rename = "type")]
